pyproject.toml
[tool.poetry]
name = "online-bank-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
coverage = "7.6.8"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
requests = "^2.32.3"
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
types-requests = "^2.32.0.20241016"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
mypy = "^1.12.1"
black = "^24.10.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
line-length = 119
exclude = '.git'

[tool.isort]
line_length = 119